type: data preparation
Added the next time-step which will be used as the output label for the
pressure. Inside the notebook, wrote the code to create the dataframe
with the simpler version to be used in the machine learning algorithm.
This simpler version considers as output only the values of the varibles
inside the cell and the neighbours.