type: Prototyping
Tried to predict the pressure field of a RANS simulation to a LES
simulation using Decision Trees and Random Forests. The results didn't
generalize, the algorithm was overfitting the data. Since the
simulation used for calibration were too transient, the velocity fields
were too different. For adjusting the results of RANS into LES it would
be necessary to evaluate somewhat stationary results, which was not the
case.
The new idea now is try to speed up the LES and RANS simulations, the
same way as was done in the laminar simulation.